[back] feat: layer parameter 타입 변경
float -> double

S11P31A107-146

+++ back/scv/src/main/java/com/scv/domain/model/dto/layer/nonlinear/ELUDTO.java
@@ -4,7 +4,7 @@ import com.scv.domain.model.dto.layer.LayerDTO;
 
 public record ELUDTO(
         String type,
-        float alpha
+        double alpha
 ) implements LayerDTO {
     @Override
     public String getType() {
