+++ back/scv/src/main/java/com/scv/domain/model/dto/layer/nonlinear/PReLUDTO.java
@@ -5,7 +5,7 @@ import com.scv.domain.model.dto.layer.LayerDTO;
 public record PReLUDTO(
         String type,
         int numParameters,
-        float init
+        double init
 ) implements LayerDTO {
     @Override
     public String getType() {
